Output is now a subclass of an Interface baseclass
